user-creation-api converted to work alongside config, discovery and gateway server
