--- UserCreationAPI/src/main/java/com/qa/cv_manager/userapi/UserapiApplication.java
@@ -1,22 +1,0 @@
-package com.qa.cv_manager.userapi;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
-import org.springframework.context.annotation.Bean;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.security.crypto.password.PasswordEncoder;
-
-@SpringBootApplication
-@EnableDiscoveryClient
-public class UserapiApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(UserapiApplication.class, args);
-	}
-	
-	@Bean
-	public PasswordEncoder passwordEncoder() {
-		return new BCryptPasswordEncoder();
-	}
-}
