--- UserCreationAPI/src/main/java/com/qa/cv_manager/userapi/persistence/domain/UserRole.java
@@ -1,42 +1,0 @@
-package com.qa.cv_manager.userapi.persistence.domain;
-
-import javax.persistence.Entity;
-import javax.persistence.Id;
-import javax.persistence.Table;
-
-@Entity
-@Table(name = "user_roles")
-public class UserRole {
-
-	@Id
-	private String username;
-	
-	private String role;
-
-	public UserRole() {}
-	
-	public UserRole(String username, String role) {
-		this.username = username;
-		this.role = role;
-	}
-	
-	public String getUsername() {
-		return username;
-	}
-	
-	public void setUsername(String username) {
-		this.username = username;
-	}
-	
-	public String getRole() {
-		return role;
-	}
-
-	public void setRole(String role) {
-		this.role = role;
-	}
-	
-	public String toString() {
-		return username + role;
-	}
-}
