--- UserCreationAPI/src/main/resources/application.properties
@@ -1,16 +1,0 @@
-server.port=8083
-
-spring.application.name=user-creation-api
-
-spring.jpa.hibernate.ddl-auto=update
-spring.datasource.url=jdbc:mysql://localhost:3306/userbase
-spring.datasource.username=root
-spring.datasource.password=root
-
-path.addUser=/addUser
-path.updatePassword=/updatePassword/{username}
-path.deleteUser=/deleteUser/{username}
-
-userbase.table.users=users
-userbase.table.roles=user_roles
-userbase.field.username=username
