+++ config/src/main/resources/application.properties
@@ -7,6 +7,6 @@ spring.cloud.config.server.git.uri=file:///${user.home}/Documents/workspace-sts-
 # ----------------------------
 # eureka connection settings
 # ----------------------------
-#eureka.client.region = default
-#eureka.client.registryFetchIntervalSeconds = 5
-#eureka.client.serviceUrl.defaultZone=http://localhost:8082/eureka/
+eureka.client.region = default
+eureka.client.registryFetchIntervalSeconds = 5
+eureka.client.serviceUrl.defaultZone=http://localhost:8082/eureka/
