--- discovery/src/main/resources/application.properties
@@ -1,2 +1,0 @@
-spring.cloud.config.name=discovery
-spring.cloud.config.uri=http://localhost:8081
