+++ user-creation-api/src/main/java/com/qa/cv_manager/usercreationapi/UserCreationApiApplication.java
@@ -1,0 +1,22 @@
+package com.qa.cv_manager.usercreationapi;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.netflix.eureka.EnableEurekaClient;
+import org.springframework.context.annotation.Bean;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.springframework.security.crypto.password.PasswordEncoder;
+
+@SpringBootApplication
+@EnableEurekaClient
+public class UserCreationApiApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(UserCreationApiApplication.class, args);
+	}
+
+	@Bean
+	public PasswordEncoder passwordEncoder() {
+		return new BCryptPasswordEncoder();
+	}
+}
