+++ user-creation-api/src/main/java/com/qa/cv_manager/usercreationapi/rest/UserRest.java
@@ -1,0 +1,37 @@
+package com.qa.cv_manager.usercreationapi.rest;
+
+import javax.validation.Valid;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RestController;
+
+import com.qa.cv_manager.usercreationapi.persistence.domain.UserPOJO;
+import com.qa.cv_manager.usercreationapi.service.UserService;
+
+@RestController
+public class UserRest {
+
+	@Autowired
+	private UserService service;
+	
+	@PostMapping("${path.addUser}")
+	public ResponseEntity<Object> addUser(@RequestBody @Valid UserPOJO user) {
+		return service.addUser(user);
+	}
+	
+	@PutMapping("${path.updatePassword}")
+	public ResponseEntity<Object> updatePassword(@RequestBody @Valid UserPOJO user, @PathVariable String username) {
+		return service.updatePassword(user, username);
+	}
+	
+	@DeleteMapping("${path.deleteUser}")
+	public ResponseEntity<Object> deleteUser(@PathVariable String username) {
+		return service.deleteUser(username);
+	}
+}
