+++ user-creation-api/src/main/java/com/qa/cv_manager/usercreationapi/service/UserService.java
@@ -1,0 +1,15 @@
+package com.qa.cv_manager.usercreationapi.service;
+
+import org.springframework.http.ResponseEntity;
+
+import com.qa.cv_manager.usercreationapi.persistence.domain.UserPOJO;
+
+public interface UserService {
+
+	ResponseEntity<Object> addUser(UserPOJO user);
+
+	ResponseEntity<Object> updatePassword(UserPOJO user, String username);
+
+	ResponseEntity<Object> deleteUser(String username);
+
+}
