+++ user-creation-api/src/main/resources/application.properties
@@ -1,0 +1,16 @@
+server.port=8084
+
+spring.application.name=user-creation-api
+
+spring.jpa.hibernate.ddl-auto=create-drop
+spring.datasource.url=jdbc:mysql://localhost:3306/cv_database
+spring.datasource.username=springuser
+spring.datasource.password=ThePassword
+
+path.addUser=/addUser
+path.updatePassword=/updatePassword/{username}
+path.deleteUser=/deleteUser/{username}
+
+userbase.table.users=users
+userbase.table.roles=user_roles
+userbase.field.username=username
