+++ user-creation-api/src/main/resources/bootstrap.properties
@@ -1,0 +1,5 @@
+spring.cloud.config.name=user-creation-api
+spring.cloud.config.discovery.service-id=config
+spring.cloud.config.discovery.enabled=true
+ 
+eureka.client.serviceUrl.defaultZone=http://localhost:8082/eureka/
