+++ user-creation-api/src/test/java/com/qa/cv_manager/usercreationapi/UserCreationApiApplicationTests.java
@@ -1,0 +1,17 @@
+package com.qa.cv_manager.usercreationapi;
+
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.junit4.SpringRunner;
+
+@RunWith(SpringRunner.class)
+@SpringBootTest
+public class UserCreationApiApplicationTests {
+
+	@Test
+	public void contextLoads() {
+	}
+
+}
+
